+++ yoseobara/src/main/java/com/final2/yoseobara/dto/request/KakaoUserInfoDto.java
@@ -1,7 +1,9 @@
 package com.final2.yoseobara.dto.request;
 
 import lombok.AllArgsConstructor;
+import lombok.Getter;
 
+@Getter
 @AllArgsConstructor
 public class KakaoUserInfoDto {
     private Long id;
