+++ yoseobara/src/main/java/com/final2/yoseobara/repository/MemberRepository.java
@@ -10,6 +10,8 @@ public interface MemberRepository extends JpaRepository<Member, Long> {
 
     Optional<Member> findByNickname(String nickname);
 
+    Optional<Member> findByKakaoId(Long kakaoId);
+
     Optional<Member> findById(Long Id);
 
 }
